% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentPlot.R
\name{enrichmentPlot}
\alias{enrichmentPlot}
\title{Enrichment plot.}
\usage{
enrichmentPlot(mGSZres, geneSetId)
}
\arguments{
\item{mGSZres}{Result object obtained after applying \code{mGSZ2} function.}

\item{geneSetId}{character with one ID present in mGSZres (a rowname of it).}
}
\value{
A ggplot2 object.
}
\description{
\code{enrichmentPlot} Enrichment plot of a gene set resulting from mGSZ2.
}
\details{
It plots the running enrichment score of a gene set.
}
\examples{
\dontrun{
data(dummyData);
mGSZres <- mGSZ2(dummyData$x, dummyData$y, dummyData$l);
enrichmentPlot(mGSZres, rownames(mGSZres)[[1]]);
}
}
\author{
Juan Cruz Rodriguez \email{jcrodriguez@bdmg.com.ar}
}
