% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startMGSZ2.R
\docType{methods}
\name{mGSZ2-GUI}
\alias{mGSZ2-GUI}
\alias{startMGSZ2}
\title{mGSZ2 GUI}
\usage{
startMGSZ2()
}
\description{
mGSZ2 package shiny-based graphical user interface (GUI)
\code{startMGSZ2} starts the GUI.
}
\examples{
## Start the GUI
\dontrun{
startMGSZ2();
}

}
